fetch function moved to fetch.js   added new file js/fetch.js
