+++ src/js/modal.view.js
@@ -1,5 +1,5 @@
-import { APIKEY } from './mainpage';
 import { showModal } from './movie-details-modal';
+import { fetchMovieById } from './fetch';
 
 const gallery = document.querySelector('.gallery__box');
 const template = document.querySelector('#film-template');
@@ -11,11 +11,12 @@ const renderMovieDetails = (id, movie) => {
   let getGenres = [...movie.genres].map(genre => genre.name).join(', ');
   template.innerHTML = `
  <section class="backdrop">
- <section class="movie-details-modal data-id="${id}"">
-  <article class="movie-details-modal__img" src="${baseURL}${movie.poster_path}" alt="${movie.original_title}">
-    </article>>
-           <article class="movie-details-modal__info"><h2>${movie.original_title}</h2>
-            <article class="movie-details-modal__list">
+    <section class="movie-details-modal data-id="${id}"">
+      <article class="movie-details-modal__img" src="${baseURL}${movie.poster_path}" alt="${movie.original_title}">
+        </article>
+           <article class="movie-details-modal__info">
+            <h2>${movie.original_title}</h2>
+              <article class="movie-details-modal__list">
                <ul class="list__headings">
                     <li> Vote / Votes ${movie.vote_average}/${movie.vote_count} </li>
                     <li>Popularity ${movie.popularity}</li>
@@ -28,13 +29,13 @@ const renderMovieDetails = (id, movie) => {
           <span class="list__headings"> / </span>
           <button type="button" class="list__btn">${movie.vote_count}</button>
         </li>
-       
+
       </ul>
-            
+
   </article>
             <h3>ABOUT</h3>
             <p>${movie.overview}</p>
-        
+
        <article class="movie-details-modal__btn-container">
             <button data-btn="watched" data-id="${id}" type="submit" class="btn-container__watched " id="add-watched">
                 ADD TO WATCHED
@@ -51,30 +52,6 @@ const renderMovieDetails = (id, movie) => {
   showModal();
 };
 
-//funkcja otwierająca film na click/ID przekazane jest do funkcji renderującej
-
-async function fetchMovieById(id) {
-  try {
-    const response = await fetch(`https://api.themoviedb.org/3/movie/${id}?api_key=${APIKEY}`);
-    const data = await response.json();
-
-    const genresResponse = await fetch(`https://api.themoviedb.org/3/genre/movie/list?api_key=${APIKEY}`);
-    const genresData = await genresResponse.json();
-    const genres = genresData.genres;
-
-    const movieGenres =
-      data.genre_ids && Array.isArray(data.genre_ids)
-        ? data.genre_ids.map(genreId => {
-            const genre = genres.find(g => g.id === genreId);
-            return genre.name;
-          })
-        : [];
-    console.log({ ...data, genre_ids: movieGenres });
-    return { ...data, genre_ids: movieGenres };
-  } catch (error) {
-    console.error(error);
-  }
-}
 function movieDetails(id) {
   fetchMovieById(id).then(movie => {
     renderMovieDetails(id, movie);
